requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# src/httpmon_kafka_pgsql/core/config.py: 2
PyYAML == 5.1.1

# src/httpmon_kafka_pgsql/agent/http_monitor.py: 2,5
aiohttp == 3.6.2

# src/httpmon_kafka_pgsql/agent/mon_producer.py: 1,2
# src/httpmon_kafka_pgsql/writer/mon_consumer.py: 2,3
aiokafka == 0.5.2

# src/httpmon_kafka_pgsql/writer/mon_pg.py: 2
asyncpg == 0.20.1
